Move control unit related apis into ControlUnit feature
