+++ frontend/src/features/ControlUnit/components/ControlUnitDialog/index.tsx
@@ -7,14 +7,14 @@ import styled from 'styled-components'
 import { AreaNote } from './AreaNote'
 import { ControlUnitContactList } from './ControlUnitContactList'
 import { ControlUnitResourceList } from './ControlUnitResourceList'
-import { RTK_DEFAULT_QUERY_OPTIONS } from '../../../../api/constants'
-import { monitorenvControlUnitApi, useGetControlUnitQuery } from '../../../../api/controlUnit'
+import { RTK_COMMON_QUERY_OPTIONS } from '../../../../api/constants'
 import { displayedComponentActions } from '../../../../domain/shared_slices/DisplayedComponent'
 import { useMainAppDispatch } from '../../../../hooks/useMainAppDispatch'
 import { useMainAppSelector } from '../../../../hooks/useMainAppSelector'
 import { FrontendApiError } from '../../../../libs/FrontendApiError'
 import { FrontendError } from '../../../../libs/FrontendError'
 import { NoRsuiteOverrideWrapper } from '../../../../ui/NoRsuiteOverrideWrapper'
+import { monitorenvControlUnitApi, useGetControlUnitQuery } from '../../controlUnitApi'
 
 export function ControlUnitDialog() {
   const dispatch = useMainAppDispatch()
@@ -25,7 +25,7 @@ export function ControlUnitDialog() {
 
   const { data: controlUnit, error: getControlControlUnitError } = useGetControlUnitQuery(
     controlUnitId,
-    RTK_DEFAULT_QUERY_OPTIONS
+    RTK_COMMON_QUERY_OPTIONS
   )
   FrontendApiError.handleIfAny(getControlControlUnitError)
