+++ frontend/src/features/ControlUnit/components/ControlUnitListDialog/FilterBar.tsx
@@ -11,18 +11,24 @@ import { useCallback, useMemo } from 'react'
 import styled from 'styled-components'
 
 import { controlUnitListDialogActions } from './slice'
-import { useGetAdministrationsQuery } from '../../../../api/administration'
-import { RTK_DEFAULT_QUERY_OPTIONS } from '../../../../api/constants'
+import { RTK_COMMON_QUERY_OPTIONS } from '../../../../api/constants'
 import { useGetStationsQuery } from '../../../../api/station'
 import { useMainAppDispatch } from '../../../../hooks/useMainAppDispatch'
 import { useMainAppSelector } from '../../../../hooks/useMainAppSelector'
+import { FrontendApiError } from '../../../../libs/FrontendApiError'
 import { isNotArchived } from '../../../../utils/isNotArchived'
+import { useGetAdministrationsQuery } from '../../administrationApi'
 
 export function FilterBar() {
   const dispatch = useMainAppDispatch()
   const filtersState = useMainAppSelector(store => store.controlUnitListDialog.filtersState)
-  const { data: administrations } = useGetAdministrationsQuery(undefined, RTK_DEFAULT_QUERY_OPTIONS)
-  const { data: bases } = useGetStationsQuery(undefined, RTK_DEFAULT_QUERY_OPTIONS)
+  const { data: administrations, error: getAdministrationsError } = useGetAdministrationsQuery(
+    undefined,
+    RTK_COMMON_QUERY_OPTIONS
+  )
+  FrontendApiError.handleIfAny(getAdministrationsError)
+  const { data: bases, error: getStationsError } = useGetStationsQuery(undefined, RTK_COMMON_QUERY_OPTIONS)
+  FrontendApiError.handleIfAny(getStationsError)
 
   const administrationsAsOptions = useMemo(
     () => getOptionsFromIdAndName((administrations || []).filter(isNotArchived)),
